Renamed StepsCollection to StepCollection
